Changed file structures, renamed directories
