+++ NetScoreCalculation/MetricCalculation/RampUp.rs
